requirements.txt
Django
django-sslserver
channels
channels-redis
python-dotenv
daphne
aif360
plotly
asgiref
cachetools
certifi
chardet
charset-normalizer
ffmpeg-python
gunicorn
google-genai
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
hume
hume["microphone"]
idna
mysqlclient
oauthlib
pillow
pip
proto-plus
protobuf
pyasn1
pyasn1_modules
pyparsing
reportlab
python-decouple
requests
requests-oauthlib
rsa
setuptools
sqlparse
tzdata
tensorflow
torch
uritemplate
urllib3
websockets
whitenoise
